+++ extensions/apple-reminders/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Apple Reminders Changelog
 
+## [Separate menu bar count from the view] - 2024-04-12
+
+- It's now possible to set reminders count type separately from the view type in menu bar.
+
 ## [Quick Add Reminder Improvements] - 2024-04-11
 
 - It's now possible to specify a due date and a list (with an `#` prefix) in `Quick Add Reminder` when not using the AI.